requirements.txt
openai
requests
pillow
python-dotenv